+++ changelog.md
@@ -6,10 +6,13 @@
 
  * db_extensions -> db_constraints
    + the repo name was changed and all followed.
+ * db_constraints.keyed.keyeditem.UniqueConstraintStructNames -> db_constraints.utils.generickey.UniqueConstraintStructNames
+ * db_constraints.keyed.keyeditem.getColumns -> db_constraints.utils.generickey.GetMembersWithUDA
 
 ### New additions
 
  * NotNull attribute
+ * db_constraints.utils.generickey.HasMembersWithUDA
 
 ### Misc.